loader: move ctypes hooks into a module
Move the ctypes hooks from bootstrap script into separate module
(pymod04_ctypes). The main motivation for this is to prevent the
modifications to global namespace from affecting the ctypes hooks
code. Specifically, if user decides to use `sys` or `os` as
variables in the global namespace, that should not effect the
ctypes hook function _frozen_name(), which treats `sys` and `os`
as names of modules (which were bound to those names when the
bootstrap script was executed).

Fixes #5797.

This commit is a cleaned-up and modernized version of corresponding
commit from #3038 (daf60a6d684f54bcc2bc2aae1bb7b1a6387deb28).
